# sumi-daemon/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sumi-daemon | v1",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:27381/"
    }
  ],
  "paths": {
...
    "/api/v1/folder/create": {
      "post": {
        "tags": [
          "folder"
        ],
        "summary": "`POST /api/v1/folder/create`",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/folder/delete": {
      "post": {
        "tags": [
          "folder"
        ],
        "summary": "`POST /api/v1/folder/delete`：目录整体移入回收站；目录已不存在时幂等清理索引残留",
        "operationId": "delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderPath"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/folder/restore": {
      "post": {
        "tags": [
          "folder"
        ],
        "summary": "`POST /api/v1/folder/restore`：从回收站恢复目录",
        "operationId": "restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderPath"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/folder/update": {
      "post": {
        "tags": [
          "folder"
        ],
        "summary": "`POST /api/v1/folder/update`：重命名/移动真实目录（索引位置与注册表键级联跟随）",
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/item/restore": {
      "post": {
        "tags": [
          "item"
        ],
        "summary": "`POST /api/v1/item/restore`：从回收站恢复（不带 path 恢复全部回收站位置）",
        "operationId": "restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteQuery": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "FolderCreate": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "parent_path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "FolderPath": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
      "FolderUpdate": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          }
        }
      },
...
    }
  }
}